POC_RAGAS/requirements.txt
ragas>=0.4.3
openai>=1.0.0
pandas>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
tabulate>=0.9.0
rapidfuzz>=3.9.0
python-dotenv>=1.0.0
